@@ -11,34 +11,45 @@ dependencies:
     sdk: flutter
 
   cupertino_icons: ^1.0.8
-  flutter_map: ^7.0.2
   latlong2: ^0.9.1
   http: ^1.2.2
-  flutter_keyboard_visibility: ^5.2.0
+  flutter_keyboard_visibility: ^6.0.0
   geolocator: ^13.0.1
   custom_info_window: ^1.0.1
   animated_text_kit: ^4.2.2
-  lottie: ^2.7.0
-  intl: ^0.17.0
+  lottie: ^3.3.1
+  intl: ^0.20.2
   flutter_dotenv: ^5.0.2
-  google_maps_flutter: ^2.5.3
+  google_maps_flutter: ^2.10.0
+  flutter_map: ^8.1.1
   shared_preferences: ^2.3.3
-  firebase_database: ^10.4.0
-  firebase_core: ^2.32.0
+  firebase_database: ^11.3.4
+  firebase_core: ^3.12.1
   firebase_core_web: ^2.18.1
-  cloud_firestore: ^4.17.5
-  flutter_typeahead: ^4.0.0
-  firebase_auth: ^4.7.0
+  cloud_firestore: ^5.6.5
+  flutter_typeahead: ^5.2.0
+  firebase_auth: ^5.5.1
   flutter_svg: ^2.0.5
+  overlay_support: ^2.1.0
+  firebase_messaging: ^15.2.4
+  flutter_local_notifications: ^19.0.0
+  url_launcher: ^6.3.1
+  flutter_animate: ^4.2.0
+  cached_network_image: ^3.3.0
+  razorpay_flutter: ^1.3.7
+  path_provider: ^2.0.15
+  pdf: ^3.10.4
+  open_file: ^3.5.10
 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^4.0.0
+  flutter_lints: ^5.0.0
 
 flutter:
   uses-material-design: true
 
   assets:
     - assets/images/
+    - .env
