@@ -1,18 +1,27 @@
+import 'dart:ffi';
+
+import 'package:final_project/UserReviewsPage.dart';
 import 'package:flutter/material.dart';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
-import 'hotels_page.dart';
-import 'transportation_page.dart';
-import 'special_places_page.dart';
-import 'map.dart';
 import 'dart:math' as math;
 import 'package:intl/intl.dart';
 import 'ActivityScreen.dart';
+import 'transportation.dart';
+import 'Map.dart';
 
 class BeachDetailPage extends StatefulWidget {
   final Map<String, dynamic> beach;
+  final double? distance;
+  final double? latitude;
+  final double? longitude;
 
-  const BeachDetailPage({Key? key, required this.beach}) : super(key: key);
+  const BeachDetailPage(
+      {super.key,
+      required this.beach,
+      this.distance,
+      this.latitude,
+      this.longitude});
 
   @override
   _BeachDetailPageState createState() => _BeachDetailPageState();
@@ -22,7 +31,7 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
   Map<String, dynamic>? currentWeatherData;
   Map<String, dynamic>? hourlyWeatherData;
   Map<String, dynamic>? openWeatherData;
-  List<Map<String, dynamic>>? dailyForecast;  // Add this state variable
+  List<Map<String, dynamic>>? dailyForecast;
   bool isLoading = true;
   String? errorMessage;
   final String openWeatherApiKey = '380f110a19b8728fdc159ab69547cbc0';
@@ -55,6 +64,42 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
       }
     }
   }
+
+  Future<void> fetchOpenWeatherData() async {
+    if (!mounted) return;
+
+    try {
+      final coordinates = widget.beach['coordinates'] as List;
+      final latitude = coordinates[0];
+      final longitude = coordinates[1];
+
+      final openWeatherUrl =
+          'https://api.openweathermap.org/data/2.5/forecast?lat=$latitude&lon=$longitude&units=metric&appid=$openWeatherApiKey';
+
+      final response = await http.get(Uri.parse(openWeatherUrl));
+
+      if (!mounted) return;
+
+      if (response.statusCode == 200) {
+        setState(() {
+          openWeatherData = json.decode(response.body);
+          isLoading = false;
+          errorMessage = null;
+        });
+      } else {
+        throw Exception('Failed to load OpenWeather data');
+      }
+    } catch (e) {
+      if (mounted) {
+        setState(() {
+          errorMessage = "Error loading OpenWeather data: ${e.toString()}";
+          isLoading = false;
+        });
+      }
+      print("Error fetching OpenWeather data: $e");
+    }
+  }
+
   void _processDailyForecast() {
     if (openWeatherData == null || !mounted) return;
 
@@ -73,6 +118,7 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
             'maxTemp': double.negativeInfinity,
             'weatherId': hourlyData['weather'][0]['id'],
             'description': hourlyData['weather'][0]['description'],
+            'icon': hourlyData['weather'][0]['icon'],
             'humidity': hourlyData['main']['humidity'],
             'windSpeed': hourlyData['wind']['speed'],
           };
@@ -80,7 +126,8 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
 
         // Safely convert temperature to double
         final dynamic tempValue = hourlyData['main']['temp'];
-        final double temp = tempValue is int ? tempValue.toDouble() : tempValue as double;
+        final double temp =
+            tempValue is int ? tempValue.toDouble() : tempValue as double;
 
         dailyMap[dateKey]!['minTemp'] = math.min(
           dailyMap[dateKey]!['minTemp'] as double,
@@ -99,6 +146,7 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
       print("Error processing daily forecast: $e");
     }
   }
+
   Future<void> fetchWeatherData() async {
     if (!mounted) return;
 
@@ -117,7 +165,8 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
 
       if (!mounted) return;
 
-      if (currentResponse.statusCode == 200 && hourlyResponse.statusCode == 200) {
+      if (currentResponse.statusCode == 200 &&
+          hourlyResponse.statusCode == 200) {
         final currentData = json.decode(currentResponse.body);
         final hourlyData = json.decode(hourlyResponse.body);
 
@@ -155,74 +204,172 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
       };
     }
 
+    // Extract weather parameters
     final temperature = currentWeatherData!['temperature'] as double;
     final windSpeed = currentWeatherData!['windspeed'] as double;
-    final precipitation = hourlyWeatherData?['precipitation'][0] as double? ?? 0.0;
-    final weatherCode = currentWeatherData!['weathercode'] as int;
-
+    final precipitation =
+        hourlyWeatherData?['precipitation'][0] as double? ?? 0.0;
     final openWeatherMain = openWeatherData!['list'][0]['main'];
     final humidity = openWeatherMain['humidity'] as int;
-    final dynamic feelsLikeValue = openWeatherMain['feels_like'];
-    final double feelsLike = feelsLikeValue is int ? feelsLikeValue.toDouble() : feelsLikeValue;
 
-    bool isDangerousWeather = weatherCode >= 95;
-    bool isHighTemperature = temperature > 35 || feelsLike > 38;
-    bool isLowTemperature = temperature < 18;
-    bool isHighWind = windSpeed > 30;
-    bool isHeavyRain = precipitation > 5;
-    bool isHighHumidity = humidity > 85;
-
-    if (isDangerousWeather) {
-      return {
-        'status': 'Unsafe',
-        'message': 'Beach visit not recommended due to thunderstorm conditions',
-        'color': Colors.red,
-        'icon': Icons.warning
-      };
-    } else if (isHighTemperature || isHighWind || isHeavyRain || isHighHumidity) {
-      return {
-        'status': 'Dangerous',
-        'message': 'Beach conditions are risky. Exercise extreme caution',
-        'color': Colors.red,
-        'icon': Icons.warning
-      };
-    } else if (isLowTemperature || windSpeed > 20 || precipitation > 2.5) {
-      return {
-        'status': 'Caution',
-        'message': 'Beach conditions require caution',
-        'color': Colors.orange,
-        'icon': Icons.warning_amber
-      };
-    } else if (windSpeed > 15 || precipitation > 1 || temperature < 22) {
-      return {
-        'status': 'Moderate',
-        'message': 'Beach conditions are moderate',
-        'color': Colors.yellow,
-        'icon': Icons.info_outline
-      };
-    } else {
-      return {
-        'status': 'Safe',
-        'message': 'Beach conditions are ideal for visiting',
-        'color': Colors.green,
-        'icon': Icons.check_circle
-      };
+    // Safety evaluation based on each parameter
+    // Temperature safety
+    String tempSafety = 'safe';
+    if ((temperature >= 24 && temperature <= 30)) {
+      tempSafety = 'safe';
+    } else if ((temperature >= 20 && temperature <= 23) ||
+        (temperature >= 31 && temperature <= 33)) {
+      tempSafety = 'moderate';
+    } else if ((temperature >= 18 && temperature <= 19) ||
+        (temperature >= 34 && temperature <= 35)) {
+      tempSafety = 'cautious';
+    } else if (temperature < 18 || temperature > 35) {
+      tempSafety = 'unsafe';
     }
+
+    // Wind speed safety
+    String windSafety = 'safe';
+    if (windSpeed <= 15) {
+      windSafety = 'safe';
+    } else if (windSpeed > 15 && windSpeed <= 30) {
+      windSafety = 'moderate';
+    } else if (windSpeed > 30 && windSpeed <= 50) {
+      windSafety = 'cautious';
+    } else if (windSpeed > 50) {
+      windSafety = 'unsafe';
+    }
+
+    // Precipitation safety
+    String precipSafety = 'safe';
+    if (precipitation <= 1.0) {
+      precipSafety = 'safe';
+    } else if (precipitation > 1.0 && precipitation <= 2.5) {
+      precipSafety = 'moderate';
+    } else if (precipitation > 2.5 && precipitation <= 5.0) {
+      precipSafety = 'cautious';
+    } else if (precipitation > 5.0) {
+      precipSafety = 'unsafe';
+    }
+
+    // Humidity safety
+    String humiditySafety = 'safe';
+    if (humidity >= 40 && humidity <= 60) {
+      humiditySafety = 'safe';
+    } else if (humidity > 60 && humidity <= 75) {
+      humiditySafety = 'moderate';
+    } else if (humidity > 75 && humidity <= 85) {
+      humiditySafety = 'cautious';
+    } else if (humidity > 85 || humidity < 40) {
+      humiditySafety = 'unsafe';
+    }
+
+    // Determine overall safety rating (using the worst rating)
+    List<String> allSafetyRatings = [
+      tempSafety,
+      windSafety,
+      precipSafety,
+      humiditySafety
+    ];
+    String overallSafety = 'safe';
+
+    if (allSafetyRatings.contains('unsafe')) {
+      overallSafety = 'unsafe';
+    } else if (allSafetyRatings.contains('cautious')) {
+      overallSafety = 'cautious';
+    } else if (allSafetyRatings.contains('moderate')) {
+      overallSafety = 'moderate';
+    }
+
+    // Create safety messages based on which parameters are problematic
+    List<String> issues = [];
+    if (tempSafety != 'safe')
+      issues.add('temperature (${temperature.toStringAsFixed(1)}°C)');
+    if (windSafety != 'safe')
+      issues.add('wind speed (${windSpeed.toStringAsFixed(1)} km/h)');
+    if (precipSafety != 'safe')
+      issues.add('precipitation (${precipitation.toStringAsFixed(1)} mm/hr)');
+    if (humiditySafety != 'safe') issues.add('humidity ($humidity%)');
+
+    String message = '';
+    if (overallSafety == 'safe') {
+      message = 'Beach conditions are ideal for visiting';
+    } else {
+      message = 'Exercise caution due to: ${issues.join(", ")}';
+    }
+
+    // Map safety status to colors and icons
+    Map<String, dynamic> statusMap = {
+      'safe': {
+        'color': Colors.green,
+        'icon': Icons.check_circle,
+        'status': 'Safe',
+      },
+      'moderate': {
+        'color': Colors.yellow.shade700,
+        'icon': Icons.info_outline,
+        'status': 'Moderately Safe',
+      },
+      'cautious': {
+        'color': Colors.orange,
+        'icon': Icons.warning_amber,
+        'status': 'Cautious',
+      },
+      'unsafe': {
+        'color': Colors.red,
+        'icon': Icons.warning,
+        'status': 'Unsafe',
+      }
+    };
+
+    return {
+      'status': statusMap[overallSafety]['status'],
+      'message': message,
+      'color': statusMap[overallSafety]['color'],
+      'icon': statusMap[overallSafety]['icon'],
+      'details': {
+        'temperature': {
+          'value': temperature,
+          'safety': tempSafety,
+        },
+        'windSpeed': {
+          'value': windSpeed,
+          'safety': windSafety,
+        },
+        'precipitation': {
+          'value': precipitation,
+          'safety': precipSafety,
+        },
+        'humidity': {
+          'value': humidity,
+          'safety': humiditySafety,
+        }
+      }
+    };
   }
 
   Widget _buildWeatherCard() {
     if (currentWeatherData == null || openWeatherData == null) {
-      return const SizedBox.shrink();
+      return const Card(
+        elevation: 4,
+        child: Padding(
+          padding: EdgeInsets.all(16.0),
+          child: Center(
+            child: Text("Weather data unavailable"),
+          ),
+        ),
+      );
     }
 
     final temperature = currentWeatherData!['temperature'] as double;
     final windSpeed = currentWeatherData!['windspeed'] as double;
-    final precipitation = hourlyWeatherData?['precipitation'][0] as double? ?? 0.0;
+    final precipitation =
+        hourlyWeatherData?['precipitation'][0] as double? ?? 0.0;
 
     final openWeatherMain = openWeatherData!['list'][0]['main'];
     final humidity = openWeatherMain['humidity'] as int;
-    final dynamic feelsLikeValue = openWeatherMain['feels_like'];
-    final double feelsLike = feelsLikeValue is int ? feelsLikeValue.toDouble() : feelsLikeValue;
+    // Not using this variable anymore, so you can comment or remove it
+    // final weatherDescription = openWeatherData!['list'][0]['weather'][0]['description'];
+    final weatherIcon = openWeatherData!['list'][0]['weather'][0]['icon'];
 
     final safetyStatus = getBeachSafetyStatus();
 
@@ -236,22 +383,13 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                const Text(
-                  "Current Weather",
-                  style: TextStyle(
-                    fontSize: 18,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-                Icon(
-                  _getWeatherIcon(currentWeatherData!['weathercode'] as int),
-                  size: 20,
-                  color: Colors.blue,
-                ),
-              ],
+            // Modified this Row to remove the weather description Container
+            const Text(
+              "Current Weather",
+              style: TextStyle(
+                fontSize: 18,
+                fontWeight: FontWeight.bold,
+              ),
             ),
             const SizedBox(height: 16),
             Row(
@@ -262,11 +400,12 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
                   children: [
                     Row(
                       children: [
-                        const Icon(Icons.thermostat, color: Colors.orange, size: 20),
+                        const Icon(Icons.thermostat,
+                            color: Colors.orange, size: 20),
                         const SizedBox(width: 8),
                         Text(
                           "${temperature.toStringAsFixed(1)}°C ",
-                          style: const TextStyle(fontSize: 18),
+                          style: const TextStyle(fontSize: 16),
                         ),
                       ],
                     ),
@@ -277,18 +416,19 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
                         const SizedBox(width: 8),
                         Text(
                           "${windSpeed.toStringAsFixed(1)} km/h",
-                          style: const TextStyle(fontSize: 18),
+                          style: const TextStyle(fontSize: 16),
                         ),
                       ],
                     ),
                     const SizedBox(height: 8),
                     Row(
                       children: [
-                        const Icon(Icons.water_drop, color: Colors.blue, size: 20),
+                        const Icon(Icons.water_drop,
+                            color: Colors.blue, size: 20),
                         const SizedBox(width: 8),
                         Text(
                           "${precipitation.toStringAsFixed(1)} mm/hr",
-                          style: const TextStyle(fontSize: 18),
+                          style: const TextStyle(fontSize: 16),
                         ),
                       ],
                     ),
@@ -299,26 +439,11 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
                         const SizedBox(width: 8),
                         Text(
                           "Humidity: $humidity%",
-                          style: const TextStyle(fontSize: 18),
+                          style: const TextStyle(fontSize: 16),
                         ),
                       ],
                     ),
                   ],
-                ),
-                Container(
-                  padding: const EdgeInsets.all(8),
-                  decoration: BoxDecoration(
-                    color: Colors.blue.withOpacity(0.1),
-                    borderRadius: BorderRadius.circular(8),
-                  ),
-                  child: Text(
-                    getWeatherDescription(currentWeatherData!['weathercode'] as int),
-                    style: const TextStyle(
-                      fontSize: 16,
-                      color: Colors.blue,
-                      fontWeight: FontWeight.w500,
-                    ),
-                  ),
                 ),
               ],
             ),
@@ -368,8 +493,202 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
     );
   }
 
+  Widget _buildHourlyForecast() {
+    if (openWeatherData == null) {
+      return const SizedBox.shrink();
+    }
 
-  Widget _buildDailyForecast() {
+    final hourlyForecasts = openWeatherData!['list'] as List;
+
+    return Card(
+      elevation: 6,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(16),
+      ),
+      child: Padding(
+        padding: const EdgeInsets.symmetric(vertical: 20.0, horizontal: 16.0),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            const Text(
+              "Hourly Forecast",
+              style: TextStyle(
+                fontSize: 22,
+                fontWeight: FontWeight.bold,
+                color: Colors.black87,
+              ),
+            ),
+            const SizedBox(height: 20),
+            SizedBox(
+              height: 160,
+              child: ListView.builder(
+                scrollDirection: Axis.horizontal,
+                itemCount: math.min(8, hourlyForecasts.length),
+                itemBuilder: (context, index) {
+                  final forecast = hourlyForecasts[index];
+                  final temp = forecast['main']['temp'];
+                  final double tempDouble =
+                      temp is int ? temp.toDouble() : temp as double;
+                  final weather = forecast['weather'][0];
+                  final icon = weather['icon'];
+                  final description = weather['description'];
+                  final time = DateTime.parse(forecast['dt_txt']);
+
+                  // Format time as 15:00 instead of 15:00
+                  final String formattedTime =
+                      '${time.hour.toString().padLeft(2, '0')}:00';
+
+                  // Generate a color based on the time of day
+                  final Color timeColor = _getTimeBasedColor(time.hour);
+                  final Color tempColor = _getTemperatureColor(tempDouble);
+
+                  return Container(
+                    width: 90,
+                    margin: const EdgeInsets.only(right: 14),
+                    decoration: BoxDecoration(
+                        gradient: LinearGradient(
+                          begin: Alignment.topCenter,
+                          end: Alignment.bottomCenter,
+                          colors: [
+                            timeColor.withOpacity(0.15),
+                            timeColor.withOpacity(0.05),
+                          ],
+                        ),
+                        borderRadius: BorderRadius.circular(16),
+                        border: Border.all(
+                          color: timeColor.withOpacity(0.3),
+                          width: 1.5,
+                        ),
+                        boxShadow: [
+                          BoxShadow(
+                            color: Colors.black.withOpacity(0.04),
+                            blurRadius: 4,
+                            offset: const Offset(0, 2),
+                          ),
+                        ]),
+                    padding:
+                        const EdgeInsets.symmetric(vertical: 14, horizontal: 8),
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: [
+                        Text(
+                          formattedTime,
+                          style: TextStyle(
+                            fontSize: 16,
+                            fontWeight: FontWeight.w600,
+                            color: timeColor.withOpacity(0.8),
+                          ),
+                        ),
+                        Container(
+                          padding: const EdgeInsets.all(8),
+                          decoration: BoxDecoration(
+                              color: Colors.white,
+                              shape: BoxShape.circle,
+                              boxShadow: [
+                                BoxShadow(
+                                  color: timeColor.withOpacity(0.2),
+                                  blurRadius: 6,
+                                  spreadRadius: 1,
+                                ),
+                              ]),
+                          child: Image.network(
+                            "https://openweathermap.org/img/wn/$icon@2x.png",
+                            width: 38,
+                            height: 38,
+                            errorBuilder: (context, error, stackTrace) {
+                              return Icon(
+                                _getWeatherIcon(description),
+                                size: 38,
+                                color: _getWeatherIconBackground(description),
+                              );
+                            },
+                          ),
+                        ),
+                        Text(
+                          '${tempDouble.round()}°',
+                          style: TextStyle(
+                            fontSize: 20,
+                            fontWeight: FontWeight.bold,
+                            color: tempColor,
+                          ),
+                        ),
+                      ],
+                    ),
+                  );
+                },
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Color _getTimeBasedColor(int hour) {
+    if (hour >= 5 && hour < 10) {
+      return Colors.orange.shade700; // Morning
+    } else if (hour >= 10 && hour < 16) {
+      return Colors.blue.shade500; // Day
+    } else if (hour >= 16 && hour < 20) {
+      return Colors.amber.shade600; // Evening
+    } else {
+      return Colors.indigo.shade600; // Night
+    }
+  }
+
+  Color _getTemperatureColor(double temp) {
+    if (temp >= 30) {
+      return Colors.deepOrange.shade700;
+    } else if (temp >= 25) {
+      return Colors.orange.shade600;
+    } else if (temp >= 20) {
+      return Colors.amber.shade600;
+    } else if (temp >= 15) {
+      return Colors.green.shade600;
+    } else if (temp >= 10) {
+      return Colors.cyan.shade600;
+    } else {
+      return Colors.blue.shade600;
+    }
+  }
+
+  Color _getWeatherIconBackground(String description) {
+    description = description.toLowerCase();
+    if (description.contains('cloud')) {
+      return Colors.blueGrey.shade400;
+    } else if (description.contains('rain') ||
+        description.contains('drizzle')) {
+      return Colors.lightBlue.shade400;
+    } else if (description.contains('thunder')) {
+      return Colors.deepPurple.shade400;
+    } else if (description.contains('snow')) {
+      return Colors.cyan.shade400;
+    } else if (description.contains('clear')) {
+      return Colors.amber.shade400;
+    } else {
+      return Colors.grey.shade400;
+    }
+  }
+
+  IconData _getWeatherIcon(String description) {
+    description = description.toLowerCase();
+    if (description.contains('cloud')) {
+      return Icons.cloud;
+    } else if (description.contains('rain') ||
+        description.contains('drizzle')) {
+      return Icons.grain;
+    } else if (description.contains('thunder')) {
+      return Icons.flash_on;
+    } else if (description.contains('snow')) {
+      return Icons.ac_unit;
+    } else if (description.contains('clear')) {
+      return Icons.wb_sunny;
+    } else {
+      return Icons.cloud;
+    }
+  }
+
+  Widget _build5DayForecast() {
     if (dailyForecast == null || dailyForecast!.isEmpty) {
       return const SizedBox.shrink();
     }
@@ -391,304 +710,142 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
                 fontWeight: FontWeight.bold,
               ),
             ),
-            const SizedBox(height: 8), // Reduced spacing here
-            ListView.builder(
-              shrinkWrap: true,
-              physics: const NeverScrollableScrollPhysics(),
-              padding: EdgeInsets.zero, // Remove default padding
-              itemCount: math.min(5, dailyForecast!.length),
-              itemBuilder: (context, index) {
-                final forecast = dailyForecast![index];
-                final date = forecast['date'] as DateTime;
-                final minTemp = forecast['minTemp'] as double;
-                final maxTemp = forecast['maxTemp'] as double;
-                final weatherId = forecast['weatherId'] as int;
-                final humidity = forecast['humidity'] as int;
-                final windSpeed = forecast['windSpeed'] as double;
+            const SizedBox(height: 16),
+            ...dailyForecast!.take(5).map((day) {
+              final date = day['date'] as DateTime;
+              final minTemp = day['minTemp'] as double;
+              final maxTemp = day['maxTemp'] as double;
+              final description = day['description'] as String;
+              final iconCode = day['icon'] as String;
+              final backgroundColor = _getWeatherIconBackground(description);
 
-                return Padding(
-                  padding: EdgeInsets.only(
-                    bottom: index != math.min(4, dailyForecast!.length - 1) ? 12.0 : 0, // Remove padding for last item
-                  ),
-                  child: Row(
-                    children: [
-                      Expanded(
-                        flex: 2,
-                        child: Text(
-                          index == 0
-                              ? 'Today'
-                              : DateFormat('EEE, MMM d').format(date),
-                          style: const TextStyle(
-                            fontSize: 14,
-                            fontWeight: FontWeight.w500,
-                          ),
-                        ),
-                      ),
-                      Expanded(
-                        flex: 1,
-                        child: Icon(
-                          _getOpenWeatherIcon(weatherId),
-                          color: Colors.blue,
-                        ),
-                      ),
-                      Expanded(
-                        flex: 2,
-                        child: Row(
-                          mainAxisAlignment: MainAxisAlignment.end,
-                          children: [
-                            Text(
-                              '${maxTemp.round()}°',
-                              style: const TextStyle(
-                                fontSize: 16,
-                                fontWeight: FontWeight.bold,
-                              ),
-                            ),
-                            Text(
-                              '${minTemp.round()}°',
-                              style: TextStyle(
-                                fontSize: 16,
-                                color: Colors.grey[600],
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                      Expanded(
-                          flex: 2,
-                          child: Row(
-                            mainAxisAlignment: MainAxisAlignment.end,
-                            children: [
-                              Icon(Icons.water_drop, size: 16, color: Colors.blue),
-                              Text(
-                                ' ${humidity}%',
-                              ),
-                            ],
-                          )
-                      ),
+              return Container(
+                margin: const EdgeInsets.only(bottom: 12),
+                padding:
+                    const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
+                decoration: BoxDecoration(
+                  borderRadius: BorderRadius.circular(10),
+                  gradient: LinearGradient(
+                    begin: Alignment.centerLeft,
+                    end: Alignment.centerRight,
+                    colors: [
+                      Colors.white,
+                      Colors.blue.withOpacity(0.05),
                     ],
                   ),
-                );
-              },
-            ),
-          ],
-        ),
-      ),
-    );
-  }  Future<void> fetchOpenWeatherData() async {
-    if (!mounted) return;
-
-    try {
-      final coordinates = widget.beach['coordinates'] as List;
-      final latitude = coordinates[0];
-      final longitude = coordinates[1];
-
-      final url =
-          'https://api.openweathermap.org/data/2.5/forecast?lat=$latitude&lon=$longitude&appid=$openWeatherApiKey&units=metric';
-
-      final response = await http.get(Uri.parse(url));
-
-      if (!mounted) return;
-
-      if (response.statusCode == 200) {
-        final data = json.decode(response.body);
-        setState(() {
-          openWeatherData = data;
-          isLoading = false;
-          errorMessage = null;
-        });
-      } else {
-        throw Exception(
-            'Failed to load OpenWeather data: ${response.statusCode}');
-      }
-    } catch (e) {
-      if (mounted) {
-        setState(() {
-          errorMessage = "Error loading OpenWeather data: ${e.toString()}";
-          isLoading = false;
-        });
-      }
-      print("Error fetching OpenWeather data: $e");
-    }
-  }
-
-
-  Widget _buildHourlyForecast() {
-    if (openWeatherData == null) {
-      return const SizedBox.shrink();
-    }
-
-    final hourlyForecasts = openWeatherData!['list'] as List;
-
-    return Card(
-      elevation: 4,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(12),
-      ),
-      child: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            const Text(
-              "Hourly Forecast",
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            const SizedBox(height: 16),
-            SizedBox(
-              height: 120,
-              child: ListView.builder(
-                scrollDirection: Axis.horizontal,
-                itemCount: math.min(8, hourlyForecasts.length),
-                itemBuilder: (context, index) {
-                  final forecast = hourlyForecasts[index];
-                  final temp = forecast['main']['temp'] as double;
-                  final weather = forecast['weather'][0];
-                  final time = DateTime.parse(forecast['dt_txt']);
-
-                  return Container(
-                    width: 80,
-                    margin: const EdgeInsets.only(right: 12),
-                    child: Column(
-                      children: [
-                        Text(
-                          '${time.hour}:00',
-                          style: const TextStyle(
-                            fontSize: 16,
-                            fontWeight: FontWeight.w500,
-                          ),
+                  border: Border.all(
+                    color: Colors.blue.withOpacity(0.1),
+                    width: 1,
+                  ),
+                ),
+                child: Row(
+                  children: [
+                    SizedBox(
+                      width: 100,
+                      child: Text(
+                        DateFormat('E, MMM d').format(date),
+                        style: const TextStyle(
+                          fontWeight: FontWeight.w600,
+                          fontSize: 15,
                         ),
-                        const SizedBox(height: 8),
-                        Icon(
-                          _getOpenWeatherIcon(weather['id'] as int),
-                          color: Colors.blue,
-                        ),
-                        const SizedBox(height: 8),
-                        Text(
-                          '${temp.round()}°C',
-                          style: const TextStyle(
-                            fontSize: 16,
-                            fontWeight: FontWeight.w500,
-                          ),
-                        ),
-                      ],
+                      ),
                     ),
-                  );
-                },
-              ),
-            ),
+                    Container(
+                      padding: const EdgeInsets.all(6),
+                      decoration: BoxDecoration(
+                        color: backgroundColor.withOpacity(0.2),
+                        shape: BoxShape.circle,
+                        border: Border.all(
+                          color: backgroundColor.withOpacity(0.3),
+                          width: 1,
+                        ),
+                      ),
+                      child: Image.network(
+                        "https://openweathermap.org/img/wn/$iconCode.png",
+                        width: 40,
+                        height: 40,
+                        errorBuilder: (context, error, stackTrace) {
+                          return Icon(
+                            Icons.cloud,
+                            size: 40,
+                            color: backgroundColor,
+                          );
+                        },
+                      ),
+                    ),
+                    const SizedBox(width: 12),
+                    Expanded(
+                      child: Column(
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: [
+                          Text(
+                            description,
+                            style: TextStyle(
+                              color: backgroundColor,
+                              fontSize: 14,
+                              fontWeight: FontWeight.w500,
+                            ),
+                          ),
+                          const SizedBox(height: 4),
+                          Row(
+                            children: [
+                              Container(
+                                padding: const EdgeInsets.symmetric(
+                                    horizontal: 8, vertical: 2),
+                                decoration: BoxDecoration(
+                                  color: Colors.blue.withOpacity(0.1),
+                                  borderRadius: BorderRadius.circular(4),
+                                ),
+                                child: Text(
+                                  '${minTemp.round()}°',
+                                  style: const TextStyle(
+                                    color: Colors.blue,
+                                    fontWeight: FontWeight.w600,
+                                  ),
+                                ),
+                              ),
+                              const SizedBox(width: 8),
+                              Container(
+                                padding: const EdgeInsets.symmetric(
+                                    horizontal: 8, vertical: 2),
+                                decoration: BoxDecoration(
+                                  color: Colors.orange.withOpacity(0.1),
+                                  borderRadius: BorderRadius.circular(4),
+                                ),
+                                child: Text(
+                                  '${maxTemp.round()}°',
+                                  style: const TextStyle(
+                                    color: Colors.deepOrange,
+                                    fontWeight: FontWeight.w600,
+                                  ),
+                                ),
+                              ),
+                            ],
+                          ),
+                        ],
+                      ),
+                    ),
+                  ],
+                ),
+              );
+            }),
           ],
         ),
       ),
     );
-  }
-
-  IconData _getOpenWeatherIcon(int id) {
-    if (id >= 200 && id < 300) return Icons.flash_on; // Thunderstorm
-    if (id >= 300 && id < 400) return Icons.grain; // Drizzle
-    if (id >= 500 && id < 600) return Icons.beach_access; // Rain
-    if (id >= 600 && id < 700) return Icons.ac_unit; // Snow
-    if (id >= 700 && id < 800) return Icons.cloud; // Atmosphere
-    if (id == 800) return Icons.wb_sunny; // Clear
-    if (id > 800) return Icons.cloud_queue; // Clouds
-    return Icons.question_mark;
-  }
-
-  String getWeatherDescription(int code) {
-    switch (code) {
-      case 0:
-        return 'Clear sky';
-      case 1:
-      case 2:
-      case 3:
-        return 'Partly cloudy';
-      case 45:
-      case 48:
-        return 'Foggy';
-      case 51:
-      case 53:
-      case 55:
-        return 'Drizzle';
-      case 61:
-      case 63:
-      case 65:
-        return 'Rain';
-      case 71:
-      case 73:
-      case 75:
-        return 'Snow';
-      case 77:
-        return 'Snow grains';
-      case 80:
-      case 81:
-      case 82:
-        return 'Rain showers';
-      case 85:
-      case 86:
-        return 'Snow showers';
-      case 95:
-        return 'Thunderstorm';
-      case 96:
-      case 99:
-        return 'Thunderstorm with hail';
-      default:
-        return 'Unknown';
-    }
-  }
-
-  IconData _getWeatherIcon(int code) {
-    switch (code) {
-      case 0:
-        return Icons.wb_sunny;
-      case 1:
-      case 2:
-      case 3:
-        return Icons.cloud_circle;
-      case 45:
-      case 48:
-        return Icons.cloud;
-      case 51:
-      case 53:
-      case 55:
-        return Icons.grain;
-      case 61:
-      case 63:
-      case 65:
-        return Icons.beach_access;
-      case 71:
-      case 73:
-      case 75:
-      case 77:
-        return Icons.ac_unit;
-      case 80:
-      case 81:
-      case 82:
-        return Icons.umbrella;
-      case 85:
-      case 86:
-        return Icons.umbrella;
-      case 95:
-      case 96:
-      case 99:
-        return Icons.flash_on;
-      default:
-        return Icons.question_mark;
-    }
   }
 
   Widget _buildActionButton(
       String title, IconData icon, VoidCallback onPressed) {
     return Container(
-      margin: const EdgeInsets.symmetric(vertical: 8),
+      margin: const EdgeInsets.symmetric(vertical: 4),
       width: double.infinity,
       child: ElevatedButton.icon(
         icon: Icon(icon),
         label: Text(title),
         onPressed: onPressed,
         style: ElevatedButton.styleFrom(
-          padding: const EdgeInsets.symmetric(vertical: 16),
+          padding: const EdgeInsets.symmetric(vertical: 12),
           backgroundColor: Colors.blue,
           foregroundColor: Colors.white,
           shape: RoundedRectangleBorder(
@@ -704,9 +861,31 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
     return Column(
       children: [
         _buildActionButton(
+          "User Activity",
+          Icons.local_activity_sharp,
+          () => Navigator.push(
+            context,
+            MaterialPageRoute(
+              builder: (context) => ActivityScreen(
+                beachName: widget.beach['name'],
+              ),
+            ),
+          ),
+        ),
+        _buildActionButton(
+          "User Feedback",
+          Icons.read_more,
+          () => Navigator.push(
+            context,
+            MaterialPageRoute(
+              builder: (context) => UserReviewsPage(beach: widget.beach),
+            ),
+          ),
+        ),
+        _buildActionButton(
           "View on Map",
           Icons.map,
-              () => Navigator.push(
+          () => Navigator.push(
             context,
             MaterialPageRoute(
               builder: (context) => MapPage(
@@ -717,33 +896,15 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
           ),
         ),
         _buildActionButton(
-          "View Nearby Hotels",
-          Icons.hotel,
-              () => Navigator.push(
+          "Transportation Services",
+          Icons.directions_car,
+          () => Navigator.push(
             context,
             MaterialPageRoute(
-              builder: (context) => HotelsPage(beach: widget.beach),
-            ),
-          ),
-        ),
-        _buildActionButton(
-          "Activity Manager",
-          Icons.local_activity_sharp,
-              () => Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) => ActivityScreen(    beachName: widget.beach['name'],
-              )
-            ),
-          ),
-        ),
-        _buildActionButton(
-          "Special Places Nearby",
-          Icons.place,
-              () => Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) => SpecialPlacesPage(beach: widget.beach),
+              builder: (context) => TransportationScreen(
+                beachName: widget.beach['name'],
+                beachData: widget.beach,
+              ),
             ),
           ),
         ),
@@ -839,45 +1000,6 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
               height: 1.5,
             ),
           ),
-          if (widget.beach['facilities'] != null) ...[
-            const SizedBox(height: 16),
-            const Text(
-              "Facilities",
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            const SizedBox(height: 8),
-            Wrap(
-              spacing: 8,
-              runSpacing: 8,
-              children: (widget.beach['facilities'] as List).map((facility) {
-                return Chip(
-                  label: Text(facility),
-                  backgroundColor: Colors.blue.withOpacity(0.1),
-                );
-              }).toList(),
-            ),
-          ],
-          if (widget.beach['bestTime'] != null) ...[
-            const SizedBox(height: 16),
-            const Text(
-              "Best Time to Visit",
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            const SizedBox(height: 8),
-            Text(
-              widget.beach['bestTime'],
-              style: TextStyle(
-                fontSize: 16,
-                color: Colors.grey[800],
-              ),
-            ),
-          ],
         ],
       ),
     );
@@ -899,7 +1021,7 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
                 const SizedBox(height: 16),
                 _buildHourlyForecast(),
                 const SizedBox(height: 16),
-                _buildDailyForecast(), // Add this line
+                _build5DayForecast(),
                 const SizedBox(height: 24),
                 _buildActionButtons(),
               ],
@@ -925,42 +1047,32 @@ class _BeachDetailPageState extends State<BeachDetailPage> {
       body: isLoading
           ? const Center(child: CircularProgressIndicator())
           : errorMessage != null
-          ? Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(
-              Icons.cloud_off,
-              size: 64,
-              color: Colors.grey[400],
-            ),
-            const SizedBox(height: 16),
-            Text(
-              errorMessage!,
-              style: TextStyle(
-                color: Colors.grey[600],
-                fontSize: 16,
-              ),
-            ),
-            const SizedBox(height: 24),
-            ElevatedButton.icon(
-              onPressed: _initializeData,
-              icon: const Icon(Icons.refresh),
-              label: const Text('Retry'),
-              style: ElevatedButton.styleFrom(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 24,
-                  vertical: 12,
-                ),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(8),
-                ),
-              ),
-            ),
-          ],
-        ),
-      )
-          : _buildBeachContent(),
+              ? Center(
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Icon(Icons.error_outline, size: 48, color: Colors.red),
+                      const SizedBox(height: 16),
+                      Text(
+                        errorMessage!,
+                        textAlign: TextAlign.center,
+                        style: const TextStyle(fontSize: 16),
+                      ),
+                      const SizedBox(height: 24),
+                      ElevatedButton(
+                        onPressed: () {
+                          setState(() {
+                            isLoading = true;
+                            errorMessage = null;
+                          });
+                          _initializeData();
+                        },
+                        child: const Text("Retry"),
+                      ),
+                    ],
+                  ),
+                )
+              : _buildBeachContent(),
     );
   }
 }
