@@ -1,32 +1,170 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'firebase_options.dart';
-import 'Welcome_Screen.dart';
+import 'package:overlay_support/overlay_support.dart';
+import 'package:flutter_dotenv/flutter_dotenv.dart' as dotenv;
+
+// Import your screens
+import 'Notification_Screen.dart';
 import 'home_view.dart';
 import 'SignupPage.dart';
 
-void main() async {
+Future<void> main() async {
+  // Ensure Flutter bindings are initialized
   WidgetsFlutterBinding.ensureInitialized();
-  await Firebase.initializeApp(
-    options: DefaultFirebaseOptions.currentPlatform,
+
+  // Load environment variables
+
+  // Initialize Firebase
+  await Firebase.initializeApp();
+
+  // Run the app with overlay support
+  runApp(
+    OverlaySupport.global(
+      child: MyApp(),
+    ),
   );
-  runApp(const MyApp());
 }
 
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+class MyApp extends StatefulWidget {
+  @override
+  _MyAppState createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  // Notification service
+  final NotificationService _notificationService = NotificationService();
+  bool isLoading = true;
+
+  @override
+  void initState() {
+    super.initState();
+    _initializeApp();
+  }
+
+  Future<void> _initializeApp() async {
+    try {
+      // Initialize notifications
+      await _notificationService.initializeNotifications();
+
+      // Update loading state
+      setState(() {
+        isLoading = false;
+      });
+    } catch (e) {
+      print('Initialization error: $e');
+      setState(() {
+        isLoading = false;
+      });
+    }
+  }
+
+  @override
+  void dispose() {
+    // Dispose of notification service resources
+    _notificationService.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Local Signup App',
+      title: 'Beach Safety App',
       theme: ThemeData(
         primarySwatch: Colors.teal,
       ),
-      home: const AuthWrapper(),
+      home: isLoading
+          ? const Scaffold(
+        body: Center(
+          child: CircularProgressIndicator(),
+        ),
+      )
+          : const AuthWrapper(),
       debugShowCheckedModeBanner: false,
     );
+  }
+}
+
+// Notification Service Class
+class NotificationService {
+  final FirebaseMessaging _firebaseMessaging = FirebaseMessaging.instance;
+  final FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
+  FlutterLocalNotificationsPlugin();
+
+  Future<void> initializeNotifications() async {
+    try {
+      // Request notification permissions
+      NotificationSettings settings = await _firebaseMessaging.requestPermission(
+        alert: true,
+        badge: true,
+        sound: true,
+      );
+
+      if (settings.authorizationStatus == AuthorizationStatus.authorized) {
+        // Initialize local notifications
+        _initializeLocalNotifications();
+
+        // Configure Firebase messaging
+        _configureFirebaseMessaging();
+      }
+    } catch (e) {
+      print('Notification initialization error: $e');
+    }
+  }
+
+  void _initializeLocalNotifications() {
+    const AndroidInitializationSettings initializationSettingsAndroid =
+    AndroidInitializationSettings('app_icon');
+
+    final InitializationSettings initializationSettings = InitializationSettings(
+      android: initializationSettingsAndroid,
+    );
+
+    flutterLocalNotificationsPlugin.initialize(
+      initializationSettings,
+      onDidReceiveNotificationResponse: (details) {
+        // Handle notification tap
+      },
+    );
+  }
+
+  void _configureFirebaseMessaging() {
+    // Handle foreground messages
+    FirebaseMessaging.onMessage.listen((RemoteMessage message) {
+      _showLocalNotification(message);
+    });
+
+    // Handle background messages
+    FirebaseMessaging.onMessageOpenedApp.listen((RemoteMessage message) {
+      // Handle notification tap when app is in background
+    });
+  }
+
+  void _showLocalNotification(RemoteMessage message) {
+    RemoteNotification? notification = message.notification;
+    AndroidNotification? android = message.notification?.android;
+
+    if (notification != null && android != null) {
+      flutterLocalNotificationsPlugin.show(
+        notification.hashCode,
+        notification.title,
+        notification.body,
+        NotificationDetails(
+          android: AndroidNotificationDetails(
+            'high_importance_channel',
+            'High Importance Notifications',
+            importance: Importance.max,
+            priority: Priority.high,
+          ),
+        ),
+      );
+    }
+  }
+
+  void dispose() {
+    // Clean up resources if needed
   }
 }
 
@@ -58,7 +196,7 @@ class _AuthWrapperState extends State<AuthWrapper> {
   @override
   Widget build(BuildContext context) {
     if (isLoading) {
-      return Scaffold(
+      return const Scaffold(
         body: Center(
           child: CircularProgressIndicator(),
         ),
